app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["aarizona <ar.ariz.arizona@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = {extras = ["redis"], version = "^5.4.0"}
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
django-markdownx = "^4.0.7"
django-markdownfield = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
